feat :: taskType 생성 함수들 tasks.ts로 옮기기
